+++ ginkgo/testsuite/test_suite.go
@@ -1,16 +1,48 @@
 package testsuite
 
 import (
+	"errors"
 	"io/ioutil"
 	"os"
 	"path/filepath"
 	"regexp"
+	"strings"
 )
 
 type TestSuite struct {
 	Path        string
 	PackageName string
 	IsGinkgo    bool
+	Precompiled bool
+}
+
+func PrecompiledTestSuite(path string) (TestSuite, error) {
+	info, err := os.Stat(path)
+	if err != nil {
+		return TestSuite{}, err
+	}
+
+	if info.IsDir() {
+		return TestSuite{}, errors.New("this is a directory, not a file")
+	}
+
+	if filepath.Ext(path) != ".test" {
+		return TestSuite{}, errors.New("this is not a .test binary")
+	}
+
+	if info.Mode()&0111 == 0 {
+		return TestSuite{}, errors.New("this is not executable")
+	}
+
+	dir := relPath(filepath.Dir(path))
+	packageName := strings.TrimSuffix(filepath.Base(path), filepath.Ext(path))
+
+	return TestSuite{
+		Path:        dir,
+		PackageName: packageName,
+		IsGinkgo:    true,
+		Precompiled: true,
+	}, nil
 }
 
 func SuitesInDir(dir string, recurse bool) []TestSuite {
@@ -36,14 +68,17 @@ func SuitesInDir(dir string, recurse bool) []TestSuite {
 	return suites
 }
 
-func New(dir string, files []os.FileInfo) TestSuite {
+func relPath(dir string) string {
 	dir, _ = filepath.Abs(dir)
 	cwd, _ := os.Getwd()
 	dir, _ = filepath.Rel(cwd, filepath.Clean(dir))
 	dir = "." + string(filepath.Separator) + dir
+	return dir
+}
 
+func New(dir string, files []os.FileInfo) TestSuite {
 	return TestSuite{
-		Path:        dir,
+		Path:        relPath(dir),
 		PackageName: packageNameForSuite(dir),
 		IsGinkgo:    filesHaveGinkgoSuite(dir, files),
 	}