+++ ginkgo/testsuite/testsuite_test.go
@@ -4,6 +4,7 @@ import (
 	"io/ioutil"
 	"os"
 	"path/filepath"
+
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/ginkgo/ginkgo/testsuite"
 	. "github.com/onsi/gomega"
@@ -13,13 +14,13 @@ var _ = Describe("TestSuite", func() {
 	var tmpDir string
 	var relTmpDir string
 
-	writeFile := func(folder string, filename string, content string) {
+	writeFile := func(folder string, filename string, content string, mode os.FileMode) {
 		path := filepath.Join(tmpDir, folder)
 		err := os.MkdirAll(path, 0700)
 		Ω(err).ShouldNot(HaveOccurred())
 
 		path = filepath.Join(path, filename)
-		ioutil.WriteFile(path, []byte(content), os.ModePerm)
+		ioutil.WriteFile(path, []byte(content), mode)
 	}
 
 	BeforeEach(func() {
@@ -34,23 +35,75 @@ var _ = Describe("TestSuite", func() {
 		Ω(err).ShouldNot(HaveOccurred())
 
 		//go files in the root directory (no tests)
-		writeFile("/", "main.go", "package main")
+		writeFile("/", "main.go", "package main", 0666)
 
 		//non-go files in a nested directory
-		writeFile("/redherring", "big_test.jpg", "package ginkgo")
+		writeFile("/redherring", "big_test.jpg", "package ginkgo", 0666)
 
 		//non-ginkgo tests in a nested directory
-		writeFile("/proffessorplum", "proffessorplum_test.go", `import "testing"`)
+		writeFile("/proffessorplum", "proffessorplum_test.go", `import "testing"`, 0666)
 
 		//ginkgo tests in a nested directory
-		writeFile("/colonelmustard", "colonelmustard_test.go", `import "github.com/onsi/ginkgo"`)
+		writeFile("/colonelmustard", "colonelmustard_test.go", `import "github.com/onsi/ginkgo"`, 0666)
 
 		//ginkgo tests in a deeply nested directory
-		writeFile("/colonelmustard/library", "library_test.go", `import "github.com/onsi/ginkgo"`)
+		writeFile("/colonelmustard/library", "library_test.go", `import "github.com/onsi/ginkgo"`, 0666)
+
+		//a precompiled ginkgo test
+		writeFile("/precompiled-dir", "precompiled.test", `fake-binary-file`, 0777)
+		writeFile("/precompiled-dir", "some-other-binary", `fake-binary-file`, 0777)
+		writeFile("/precompiled-dir", "nonexecutable.test", `fake-binary-file`, 0666)
 	})
 
 	AfterEach(func() {
 		os.RemoveAll(tmpDir)
+	})
+
+	Describe("Finding precompiled test suites", func() {
+		Context("if pointed at an executable file that ends with .test", func() {
+			It("should return a precompiled test suite", func() {
+				suite, err := PrecompiledTestSuite(filepath.Join(tmpDir, "precompiled-dir", "precompiled.test"))
+				Ω(err).ShouldNot(HaveOccurred())
+				Ω(suite).Should(Equal(TestSuite{
+					Path:        relTmpDir + "/precompiled-dir",
+					PackageName: "precompiled",
+					IsGinkgo:    true,
+					Precompiled: true,
+				}))
+			})
+		})
+
+		Context("if pointed at a directory", func() {
+			It("should error", func() {
+				suite, err := PrecompiledTestSuite(filepath.Join(tmpDir, "precompiled-dir"))
+				Ω(suite).Should(BeZero())
+				Ω(err).Should(HaveOccurred())
+			})
+		})
+
+		Context("if pointed at an executable that doesn't have .test", func() {
+			It("should error", func() {
+				suite, err := PrecompiledTestSuite(filepath.Join(tmpDir, "precompiled-dir", "some-other-binary"))
+				Ω(suite).Should(BeZero())
+				Ω(err).Should(HaveOccurred())
+			})
+		})
+
+		Context("if pointed at a .test that isn't executable", func() {
+			It("should error", func() {
+				suite, err := PrecompiledTestSuite(filepath.Join(tmpDir, "precompiled-dir", "nonexecutable.test"))
+				Ω(suite).Should(BeZero())
+				Ω(err).Should(HaveOccurred())
+			})
+		})
+
+		Context("if pointed at a nonexisting file", func() {
+			It("should error", func() {
+				suite, err := PrecompiledTestSuite(filepath.Join(tmpDir, "precompiled-dir", "nope-nothing-to-see-here"))
+				Ω(suite).Should(BeZero())
+				Ω(err).Should(HaveOccurred())
+			})
+		})
 	})
 
 	Describe("scanning for suites in a directory", func() {
@@ -69,6 +122,7 @@ var _ = Describe("TestSuite", func() {
 				Ω(suites[0].Path).Should(Equal(relTmpDir + "/colonelmustard"))
 				Ω(suites[0].PackageName).Should(Equal("colonelmustard"))
 				Ω(suites[0].IsGinkgo).Should(BeTrue())
+				Ω(suites[0].Precompiled).Should(BeFalse())
 			})
 		})
 
@@ -80,6 +134,7 @@ var _ = Describe("TestSuite", func() {
 				Ω(suites[0].Path).Should(Equal(relTmpDir + "/proffessorplum"))
 				Ω(suites[0].PackageName).Should(Equal("proffessorplum"))
 				Ω(suites[0].IsGinkgo).Should(BeFalse())
+				Ω(suites[0].Precompiled).Should(BeFalse())
 			})
 		})
 
@@ -92,16 +147,19 @@ var _ = Describe("TestSuite", func() {
 					Path:        relTmpDir + "/colonelmustard",
 					PackageName: "colonelmustard",
 					IsGinkgo:    true,
+					Precompiled: false,
 				}))
 				Ω(suites).Should(ContainElement(TestSuite{
 					Path:        relTmpDir + "/proffessorplum",
 					PackageName: "proffessorplum",
 					IsGinkgo:    false,
+					Precompiled: false,
 				}))
 				Ω(suites).Should(ContainElement(TestSuite{
 					Path:        relTmpDir + "/colonelmustard/library",
 					PackageName: "library",
 					IsGinkgo:    true,
+					Precompiled: false,
 				}))
 			})
 		})